// src/models/IMovieData.ts
export interface IMovieData {
  id: string
  title: string
  year: string
  genres: string[]
  ratings: number[]
  poster: string
  contentRating: string
  duration: string
  releaseDate: string
  averageRating: number
  originalTitle: string
  storyline: string
  actors: string[]
  imdbRating: string
  posterurl: string
  isFavourite: boolean
}